+++ backend/src/parsers/chatgpt_parsers.py
