+++ backend/tests/integration/test_writingprompt.py
@@ -10,7 +10,7 @@ from unittest.mock import MagicMock
 import pytest
 from fastapi.testclient import TestClient
 
-from main import app
+from backend.src.main import app
 
 
 class TestWritingPrompt(unittest.TestCase):
