+++ backend/tests/unit/gemini/test_gemini_client.py
@@ -6,7 +6,7 @@ import unittest
 from unittest import mock
 from unittest.mock import MagicMock
 
-from src.gemini.client import GoogleGeminiClient
+from backend.src.gemini.client import GoogleGeminiClient
 
 
 class TestGeminiClient(unittest.TestCase):
