+++ backend/tests/unit/handlers/test_writing_prompt_handlers.py
@@ -6,7 +6,7 @@ import unittest
 from unittest import mock
 from unittest.mock import MagicMock
 
-from src.handlers.writingprompthandlers import create_writing_prompt
+from backend.src.handlers.writingprompthandlers import create_writing_prompt
 
 
 class TestWritingPromptHandlers(unittest.TestCase):
