+++ backend/tests/unit/parsers/test_chatgpt_parsers.py
@@ -5,7 +5,7 @@ Unit tests for ChatGPT parsers
 import unittest
 from unittest.mock import MagicMock
 
-from src.parsers.chatgpt_parsers import parse_chatgpt_single_response
+from backend.src.parsers.chatgpt_parsers import parse_chatgpt_single_response
 
 
 class TestChatGPTParsers(unittest.TestCase):
